feature(HS): handle axios response and segregated files for old and new api
