+++ v0/destinations/hs/transform.js
@@ -1,406 +1,16 @@
-const get = require("get-value");
-const { EventType, MappedToDestinationKey } = require("../../../constants");
+const { EventType } = require("../../../constants");
+const { getErrorRespEvents, CustomError } = require("../../util");
 const {
-  defaultGetRequestConfig,
-  defaultPostRequestConfig,
-  defaultRequestConfig,
-  getFieldValueFromMessage,
-  getSuccessRespEvents,
-  getErrorRespEvents,
-  CustomError,
-  addExternalIdToTraits,
-  defaultBatchRequestConfig,
-  removeUndefinedAndNullValues,
-  getDestinationExternalID,
-  constructPayload,
-  isDefinedAndNotNullAndNotEmpty,
-  getDestinationExternalIDInfoForRetl
-} = require("../../util");
+  processLegacyIdentify,
+  processLegacyTrack,
+  legacyBatchEvents
+} = require("./HSTransform-v1");
 const {
-  BATCH_CONTACT_ENDPOINT,
-  MAX_BATCH_SIZE,
-  TRACK_ENDPOINT,
-  IDENTIFY_CREATE_UPDATE_CONTACT,
-  IDENTIFY_CREATE_NEW_CONTACT,
-  hsCommonConfigJson,
-  IDENTIFY_CRM_UPDATE_NEW_CONTACT,
-  IDENTIFY_CRM_CREATE_NEW_CONTACT,
-  MAX_BATCH_SIZE_CRM_CONTACT,
-  BATCH_IDENTIFY_CRM_CREATE_NEW_CONTACT,
-  BATCH_IDENTIFY_CRM_UPDATE_NEW_CONTACT,
-  mappingConfig,
-  ConfigCategory,
-  TRACK_CRM_ENDPOINT,
-  CRM_CREATE_CUSTOM_OBJECTS,
-  MAX_BATCH_SIZE_CRM_OBJECT,
-  BATCH_CREATE_CUSTOM_OBJECTS
-} = require("./config");
-const {
-  getTraits,
-  getProperties,
-  getTransformedJSON,
-  getEmailAndUpdatedProps,
-  formatPropertyValueForIdentify,
-  searchContacts,
-  getCRMUpdatedProps,
-  getEventAndPropertiesFromConfig
-} = require("./util");
-
-/**
- * Using New API
- * Ref - https://developers.hubspot.com/docs/api/crm/contacts
- * @param {*} message
- * @param {*} destination
- * @param {*} propertyMap
- * @returns
- */
-const processIdentify = async (message, destination, propertyMap) => {
-  const { Config } = destination;
-  const traits = getFieldValueFromMessage(message, "traits");
-  const mappedToDestination = get(message, MappedToDestinationKey);
-  // if mappedToDestination is set true, then add externalId to traits
-  if (mappedToDestination) {
-    addExternalIdToTraits(message);
-  } else if (!traits || !traits.email) {
-    throw new CustomError(
-      "[HS]:: Identify without email is not supported.",
-      400
-    );
-  }
-
-  // if (!traits || !traits.email) {
-  //   throw new CustomError(
-  //     "[HS]:: Identify without email is not supported.",
-  //     400
-  //   );
-  // }
-
-  let contactId = getDestinationExternalID(message, "hsContactId");
-
-  if (!contactId) {
-    contactId = await searchContacts(message, destination);
-    if (!contactId && traits.email) {
-      contactId = await searchContacts(message, destination, "email");
-    }
-  }
-
-  const userProperties = await getTransformedJSON(
-    message,
-    hsCommonConfigJson,
-    destination,
-    propertyMap
-  );
-
-  const payload = {
-    properties: formatPropertyValueForIdentify(userProperties)
-  };
-
-  // build response
-  let endpoint;
-  const response = defaultRequestConfig();
-
-  // for rETL source support for custom objects
-  // Ref - https://developers.hubspot.com/docs/api/crm/crm-custom-objects
-  if (mappedToDestination) {
-    const { objectType } = getDestinationExternalIDInfoForRetl(message, "HS");
-    endpoint = CRM_CREATE_CUSTOM_OBJECTS.replace(":objectType", objectType);
-    response.body.JSON = removeUndefinedAndNullValues({ properties: traits });
-  } else {
-    if (contactId) {
-      // update
-      endpoint = IDENTIFY_CRM_UPDATE_NEW_CONTACT.replace(
-        ":contactId",
-        contactId
-      );
-    } else {
-      // create
-      endpoint = IDENTIFY_CRM_CREATE_NEW_CONTACT;
-    }
-    response.body.JSON = removeUndefinedAndNullValues(payload);
-  }
-
-  response.endpoint = endpoint;
-  response.method = defaultPostRequestConfig.requestMethod;
-  response.headers = {
-    "Content-Type": "application/json"
-  };
-
-  // choosing API Type
-  if (Config.authorizationType === "newPrivateAppApi") {
-    // Private Apps
-    response.headers = {
-      ...response.headers,
-      Authorization: `Bearer ${Config.accessToken}`
-    };
-  } else {
-    // use legacy API Key
-    response.params = { hapikey: Config.apiKey };
-  }
-
-  return response;
-};
-
-/**
- * using legacy API
- * Reference:
- * https://legacydocs.hubspot.com/docs/methods/contacts/create_contact
- * https://legacydocs.hubspot.com/docs/methods/contacts/create_or_update
- *
- * for rETL support for custom objects
- * Ref - https://developers.hubspot.com/docs/api/crm/crm-custom-objects
- * @param {*} message
- * @param {*} destination
- * @param {*} propertyMap
- * @returns
- */
-const processLegacyIdentify = async (message, destination, propertyMap) => {
-  const { Config } = destination;
-  const traits = getFieldValueFromMessage(message, "traits");
-  const mappedToDestination = get(message, MappedToDestinationKey);
-  // if mappedToDestination is set true, then add externalId to traits
-  // rETL source
-  if (mappedToDestination) {
-    addExternalIdToTraits(message);
-  } else if (!traits || !traits.email) {
-    throw new CustomError(
-      "[HS]:: Identify without email is not supported.",
-      400
-    );
-  }
-
-  const userProperties = await getTransformedJSON(
-    message,
-    hsCommonConfigJson,
-    destination,
-    propertyMap
-  );
-
-  const payload = {
-    properties: formatPropertyValueForIdentify(userProperties)
-  };
-
-  // build response
-  const { email } = traits;
-  let endpoint;
-  const response = defaultRequestConfig();
-
-  // for rETL source support for custom objects
-  // Ref - https://developers.hubspot.com/docs/api/crm/crm-custom-objects
-  if (mappedToDestination) {
-    const { objectType } = getDestinationExternalIDInfoForRetl(message, "HS");
-    endpoint = CRM_CREATE_CUSTOM_OBJECTS.replace(":objectType", objectType);
-    response.body.JSON = removeUndefinedAndNullValues({ properties: traits });
-  } else {
-    if (email) {
-      endpoint = IDENTIFY_CREATE_UPDATE_CONTACT.replace(
-        ":contact_email",
-        email
-      );
-    } else {
-      endpoint = IDENTIFY_CREATE_NEW_CONTACT;
-    }
-    response.body.JSON = removeUndefinedAndNullValues(payload);
-  }
-
-  response.endpoint = endpoint;
-  response.method = defaultPostRequestConfig.requestMethod;
-  response.headers = {
-    "Content-Type": "application/json"
-  };
-
-  // choosing API Type
-  if (Config.authorizationType === "newPrivateAppApi") {
-    // Private Apps
-    response.headers = {
-      ...response.headers,
-      Authorization: `Bearer ${Config.accessToken}`
-    };
-  } else {
-    // use legacy API Key
-    response.params = { hapikey: Config.apiKey };
-  }
-
-  return response;
-};
-
-/**
- * CRM API
- * Associations v3
- * here we are associating objectType to contact only
- * Ref - https://developers.hubspot.com/docs/api/crm/associations/v3
- * @param {*} message
- * @param {*} destination
- * @param {*} propertyMap
- */
-// const processCRMCustomObjects = async (message, destination, traits) => {
-//   const { Config } = destination;
-//   let response = {};
-
-//   const { contactId, qualifiedName, objects } = traits.hubspot;
-//   if (!contactId) {
-//     throw new Error(
-//       "HubSpot contactId is not provided. Aborting custom-object association",
-//       400
-//     );
-//   }
-
-//   if (!qualifiedName) {
-//     throw new Error(
-//       "HubSpot qualifiedName is not provided. Aborting custom-object association",
-//       400
-//     );
-//   }
-
-//   if (!objects || !Array.isArray(objects) || objects.length === 0) {
-//     throw new Error(
-//       "HubSpot objects are not provided.  Aborting custom-object association",
-//       400
-//     );
-//   }
-
-//   const endpoint = CRM_ASSOCIATION_V3.replace(
-//     ":fromObjectType",
-//     qualifiedName
-//   ).replace(":toObjectType", "contact");
-
-//   const inputs = [];
-//   objects.forEach(item => {
-//     inputs.push({
-//       from: { id: item.objectId },
-//       to: { id: contactId },
-//       type: `${item.objectType}_to_contact`
-//     });
-//   });
-
-//   // creating response
-//   response = defaultRequestConfig();
-//   response.endpoint = endpoint;
-//   response.headers = {
-//     "Content-Type": "application/json"
-//   };
-//   response.body.JSON = { inputs };
-
-//   // choosing API Type
-//   if (Config.authorizationType === "newPrivateAppApi") {
-//     // Private Apps
-//     response.headers = {
-//       ...response.headers,
-//       Authorization: `Bearer ${Config.accessToken}`
-//     };
-//   } else {
-//     // use legacy API Key
-//     response.params = { hapikey: Config.apiKey };
-//   }
-
-//   return response;
-// };
-
-/**
- * using New API
- * Ref - https://developers.hubspot.com/docs/api/analytics/events
- * @param {*} message
- * @param {*} destination
- * @returns
- */
-const processTrack = async (message, destination) => {
-  const { Config } = destination;
-
-  let payload = constructPayload(
-    message,
-    mappingConfig[ConfigCategory.TRACK.name]
-  );
-
-  // fetch event name and its properties from config (webapp)
-  payload = getEventAndPropertiesFromConfig(message, destination, payload);
-
-  payload.properties = {
-    ...payload.properties,
-    ...constructPayload(
-      message,
-      mappingConfig[ConfigCategory.TRACK_PROPERTIES.name]
-    )
-  };
-
-  // either of email or utk or contactId should be present
-  if (!payload.email && !payload.utk && !payload.contactId) {
-    throw new CustomError(
-      "[HS]:: either of email, utk or contactId is required for custom behavioral events",
-      400
-    );
-  }
-
-  const response = defaultRequestConfig();
-  response.endpoint = TRACK_CRM_ENDPOINT;
-  response.method = defaultPostRequestConfig.requestMethod;
-  response.headers = {
-    "Content-Type": "application/json"
-  };
-  response.body.JSON = removeUndefinedAndNullValues(payload);
-
-  // choosing API Type
-  if (Config.authorizationType === "newPrivateAppApi") {
-    // remove hubId
-    // eslint-disable-next-line no-underscore-dangle
-    response.headers = {
-      ...response.headers,
-      Authorization: `Bearer ${Config.accessToken}`
-    };
-  } else {
-    // using legacyApiKey
-    response.endpoint = `${TRACK_CRM_ENDPOINT}?hapikey=${Config.hapikey}`;
-  }
-
-  return response;
-};
-
-/**
- * using legacy API
- * Ref - https://legacydocs.hubspot.com/docs/methods/enterprise_events/http_api
- * @param {*} message
- * @param {*} destination
- * @param {*} propertyMap
- * @returns
- */
-const processLegacyTrack = async (message, destination, propertyMap) => {
-  const { Config } = destination;
-  let parameters = {
-    _a: Config.hubID,
-    _n: message.event,
-    _m: get(message, "properties.revenue") || get(message, "properties.value"),
-    id: getDestinationExternalID(message, "hubspotId")
-  };
-
-  parameters = removeUndefinedAndNullValues(parameters);
-  const userProperties = await getTransformedJSON(
-    message,
-    hsCommonConfigJson,
-    destination,
-    propertyMap
-  );
-
-  const payload = { ...parameters, ...userProperties };
-  const params = removeUndefinedAndNullValues(payload);
-
-  const response = defaultRequestConfig();
-  response.endpoint = TRACK_ENDPOINT;
-  response.method = defaultGetRequestConfig.requestMethod;
-  response.headers = {
-    "Content-Type": "application/json"
-  };
-
-  // choosing API Type
-  if (Config.authorizationType === "newPrivateAppApi") {
-    // eslint-disable-next-line no-underscore-dangle
-    delete params._a;
-    response.headers = {
-      ...response.headers,
-      Authorization: `Bearer ${Config.accessToken}`
-    };
-  }
-  response.params = params;
-
-  return response;
-};
+  processIdentify,
+  processTrack,
+  batchEvents
+} = require("./HSTransform-v2");
+const { getTraits, getProperties } = require("./util");
 
 const processSingleMessage = async (message, destination, propertyMap) => {
   if (!message.type) {
@@ -451,365 +61,6 @@ const processSingleMessage = async (message, destination, propertyMap) => {
 // has been deprecated - using routerTransform for both the versions
 const process = event => {
   return processSingleMessage(event.message, event.destination);
-};
-
-const batchIdentify = (
-  arrayChunksIdentify,
-  batchedResponseList,
-  batchOperation
-) => {
-  // list of chunks [ [..], [..] ]
-  arrayChunksIdentify.forEach(chunk => {
-    const identifyResponseList = [];
-    const metadata = [];
-
-    // extracting message, destination value
-    // from the first event in a batch
-    const { message, destination } = chunk[0];
-
-    let batchEventResponse = defaultBatchRequestConfig();
-
-    if (batchOperation === "create") {
-      // create operation
-      chunk.forEach(ev => {
-        // format properties into batch structure
-        // eslint-disable-next-line no-param-reassign
-        ev.message.body.JSON.properties = getCRMUpdatedProps(
-          ev.message.body.JSON.properties
-        );
-
-        // duplicate email can cause issue with create in batch
-        // updating the existing one to avoid duplicate
-        // as same event can fire in batch one of the reason
-        // can be due to network lag or processor being busy
-        const isDuplicate = identifyResponseList.find(data => {
-          return (
-            data.properties.email === ev.message.body.JSON.properties.email
-          );
-        });
-        if (isDefinedAndNotNullAndNotEmpty(isDuplicate)) {
-          // array is being shallow copied hence changes are affecting the original reference
-          // basically rewriting the same value to avoid duplicate entry
-          isDuplicate.properties = ev.message.body.JSON.properties;
-        } else {
-          // appending unique events
-          identifyResponseList.push({
-            properties: ev.message.body.JSON.properties
-          });
-        }
-        metadata.push(ev.metadata);
-      });
-    } else if (batchOperation === "update") {
-      // update operation
-      chunk.forEach(ev => {
-        // eslint-disable-next-line no-param-reassign
-        ev.message.body.JSON.properties = getCRMUpdatedProps(
-          ev.message.body.JSON.properties
-        );
-        // update has contactId and properties
-        // extract contactId from the end of the endpoint
-        const id = ev.message.endpoint.split("/").pop();
-
-        // duplicate contactId is not allowed in batch
-        // updating the existing one to avoid duplicate
-        // as same event can fire in batch one of the reason
-        // can be due to network lag or processor being busy
-        const isDuplicate = identifyResponseList.find(data => {
-          return data.id === id;
-        });
-        if (isDefinedAndNotNullAndNotEmpty(isDuplicate)) {
-          // rewriting the same value to avoid duplicate entry
-          isDuplicate.properties = ev.message.body.JSON.properties;
-        } else {
-          // appending unique events
-          identifyResponseList.push({
-            id,
-            properties: ev.message.body.JSON.properties
-          });
-        }
-        metadata.push(ev.metadata);
-      });
-    } else if (batchOperation === "general") {
-      // general identify event
-
-      /* Note: */
-      // for now it is just HS CRM custom objects however later it can be
-      // refactored to accomodate upcoming endpoint and make necessary changes
-
-      chunk.forEach(ev => {
-        const { properties } = ev.message.body.JSON;
-        if (properties && !Array.isArray(properties)) {
-          identifyResponseList.push({ ...ev.message.body.JSON });
-          metadata.push(ev.metadata);
-        }
-      });
-
-      // CRM_CREATE_CUSTOM_OBJECTS has objectType value
-      // extract objectType from the end of the endpoint
-      const objectType = chunk[0].message.endpoint.split("/").pop();
-
-      batchEventResponse.batchedRequest.endpoint = BATCH_CREATE_CUSTOM_OBJECTS.replace(
-        ":objectType",
-        objectType
-      );
-    }
-
-    batchEventResponse.batchedRequest.body.JSON = {
-      inputs: identifyResponseList
-    };
-
-    if (batchOperation === "create") {
-      batchEventResponse.batchedRequest.endpoint = BATCH_IDENTIFY_CRM_CREATE_NEW_CONTACT;
-    } else if (batchOperation === "update") {
-      batchEventResponse.batchedRequest.endpoint = BATCH_IDENTIFY_CRM_UPDATE_NEW_CONTACT;
-    }
-
-    batchEventResponse.batchedRequest.headers = message.headers;
-    batchEventResponse.batchedRequest.params = message.params;
-
-    batchEventResponse = {
-      ...batchEventResponse,
-      metadata,
-      destination
-    };
-    batchedResponseList.push(
-      getSuccessRespEvents(
-        batchEventResponse.batchedRequest,
-        batchEventResponse.metadata,
-        batchEventResponse.destination,
-        true
-      )
-    );
-  });
-
-  return batchedResponseList;
-};
-
-const batchEvents = destEvents => {
-  let batchedResponseList = [];
-  const trackResponseList = [];
-  // create contact chunck
-  let createContactEventsChunk = [];
-  // update contact chunk
-  let updateContactEventsChunk = [];
-  // general indentify event chunk
-  let eventsChunk = [];
-  const arrayChunksIdentifyCreateContact = [];
-  const arrayChunksIdentifyUpdateContact = [];
-  const arrayChunksIdentify = [];
-  destEvents.forEach((event, index) => {
-    // handler for track call
-    // track call does not have batch endpoint
-    if (event.message.endpoint.includes("/events/v3/send")) {
-      const { message, metadata, destination } = event;
-      const endpoint = get(message, "endpoint");
-
-      const batchedResponse = defaultBatchRequestConfig();
-      batchedResponse.batchedRequest.headers = message.headers;
-      batchedResponse.batchedRequest.endpoint = endpoint;
-      batchedResponse.batchedRequest.body = message.body;
-      batchedResponse.batchedRequest.params = message.params;
-      batchedResponse.batchedRequest.method =
-        defaultGetRequestConfig.requestMethod;
-      batchedResponse.metadata = [metadata];
-      batchedResponse.destination = destination;
-
-      trackResponseList.push(
-        getSuccessRespEvents(
-          batchedResponse.batchedRequest,
-          batchedResponse.metadata,
-          batchedResponse.destination
-        )
-      );
-    } else if (
-      event.message.endpoint ===
-      "https://api.hubapi.com/crm/v3/objects/contacts"
-    ) {
-      // Identify: making chunks for CRM create contact endpoint
-      createContactEventsChunk.push(event);
-    } else if (
-      event.message.endpoint ===
-      "https://api.hubapi.com/crm/v3/objects/contacts/"
-    ) {
-      // Identify: making chunks for CRM update contact endpoint
-      updateContactEventsChunk.push(event);
-    } else {
-      eventsChunk.push(event);
-    }
-
-    // CRM create contact endpoint chunks
-    if (
-      createContactEventsChunk.length &&
-      (createContactEventsChunk.length === MAX_BATCH_SIZE_CRM_CONTACT ||
-        index === destEvents.length - 1)
-    ) {
-      arrayChunksIdentifyCreateContact.push(createContactEventsChunk);
-      createContactEventsChunk = [];
-    }
-
-    // CRM update contact endpoint chunks
-    if (
-      updateContactEventsChunk.length &&
-      (updateContactEventsChunk.length === MAX_BATCH_SIZE_CRM_CONTACT ||
-        index === destEvents.length - 1)
-    ) {
-      arrayChunksIdentifyUpdateContact.push(updateContactEventsChunk);
-      updateContactEventsChunk = [];
-    }
-
-    // general identify chunks
-    if (
-      eventsChunk.length &&
-      (eventsChunk.length === MAX_BATCH_SIZE_CRM_OBJECT ||
-        index === destEvents.length - 1)
-    ) {
-      arrayChunksIdentify.push(eventsChunk);
-      eventsChunk = [];
-    }
-  });
-
-  // batching up 'create' contact endpoint chunks
-  if (arrayChunksIdentifyCreateContact.length) {
-    batchedResponseList = batchIdentify(
-      arrayChunksIdentifyCreateContact,
-      batchedResponseList,
-      "create"
-    );
-  }
-
-  // batching up 'update' contact endpoint chunks
-  if (arrayChunksIdentifyUpdateContact.length) {
-    batchedResponseList = batchedResponseList.concat(
-      batchIdentify(
-        arrayChunksIdentifyUpdateContact,
-        batchedResponseList,
-        "update"
-      )
-    );
-  }
-
-  // batching up 'general' identify endpoint chunks
-  if (arrayChunksIdentify.length) {
-    batchedResponseList = batchedResponseList.concat(
-      batchIdentify(arrayChunksIdentify, batchedResponseList, "general")
-    );
-  }
-
-  return batchedResponseList.concat(trackResponseList);
-};
-
-const legacyBatchEvents = destEvents => {
-  const batchedResponseList = [];
-  const trackResponseList = [];
-  let eventsChunk = [];
-  const arrayChunksIdentify = [];
-  destEvents.forEach((event, index) => {
-    // handler for track call
-    if (event.message.method === "GET") {
-      const { message, metadata, destination } = event;
-      const endpoint = get(message, "endpoint");
-
-      const batchedResponse = defaultBatchRequestConfig();
-      batchedResponse.batchedRequest.headers = message.headers;
-      batchedResponse.batchedRequest.endpoint = endpoint;
-      batchedResponse.batchedRequest.body = message.body;
-      batchedResponse.batchedRequest.params = message.params;
-      batchedResponse.batchedRequest.method =
-        defaultGetRequestConfig.requestMethod;
-      batchedResponse.metadata = [metadata];
-      batchedResponse.destination = destination;
-
-      trackResponseList.push(
-        getSuccessRespEvents(
-          batchedResponse.batchedRequest,
-          batchedResponse.metadata,
-          batchedResponse.destination
-        )
-      );
-    } else {
-      // making chunks for identify
-      eventsChunk.push(event);
-    }
-    if (
-      eventsChunk.length &&
-      (eventsChunk.length === MAX_BATCH_SIZE || index === destEvents.length - 1)
-    ) {
-      arrayChunksIdentify.push(eventsChunk);
-      eventsChunk = [];
-    }
-  });
-
-  // list of chunks [ [..], [..] ]
-  arrayChunksIdentify.forEach(chunk => {
-    const identifyResponseList = [];
-    const metadata = [];
-
-    // extracting destination, apiKey value
-    // from the first event in a batch
-    const { destination } = chunk[0];
-    const { Config } = destination;
-
-    let batchEventResponse = defaultBatchRequestConfig();
-
-    chunk.forEach(ev => {
-      const { properties } = ev.message.body.JSON;
-      if (properties && !Array.isArray(properties)) {
-        identifyResponseList.push({ ...ev.message.body.JSON });
-        batchEventResponse.batchedRequest.body.JSON = {
-          inputs: identifyResponseList
-        };
-        batchEventResponse.batchedRequest.endpoint = `${ev.message.endpoint}/batch/create`;
-        metadata.push(ev.metadata);
-      } else {
-        const { email, updatedProperties } = getEmailAndUpdatedProps(
-          ev.message.body.JSON.properties
-        );
-        // eslint-disable-next-line no-param-reassign
-        ev.message.body.JSON.properties = updatedProperties;
-        identifyResponseList.push({
-          email,
-          properties: ev.message.body.JSON.properties
-        });
-        metadata.push(ev.metadata);
-        batchEventResponse.batchedRequest.body.JSON_ARRAY = {
-          batch: JSON.stringify(identifyResponseList)
-        };
-        batchEventResponse.batchedRequest.endpoint = BATCH_CONTACT_ENDPOINT;
-      }
-    });
-
-    batchEventResponse.batchedRequest.headers = {
-      "Content-Type": "application/json"
-    };
-
-    // choosing API Type
-    if (Config.authorizationType === "newPrivateAppApi") {
-      // Private Apps
-      batchEventResponse.batchedRequest.headers = {
-        ...batchEventResponse.batchedRequest.headers,
-        Authorization: `Bearer ${Config.accessToken}`
-      };
-    } else {
-      // API Key
-      batchEventResponse.batchedRequest.params = { hapikey: Config.apiKey };
-    }
-
-    batchEventResponse = {
-      ...batchEventResponse,
-      metadata,
-      destination
-    };
-    batchedResponseList.push(
-      getSuccessRespEvents(
-        batchEventResponse.batchedRequest,
-        batchEventResponse.metadata,
-        batchEventResponse.destination,
-        true
-      )
-    );
-  });
-
-  return batchedResponseList.concat(trackResponseList);
 };
 
 // we are batching by default at routerTransform
@@ -875,12 +126,13 @@ const processRouterDest = async inputs => {
     })
   );
 
+  // batch implementation
   let batchedResponseList = [];
   if (successRespList.length) {
     if (destination.Config.apiVersion === "newApi") {
-      batchedResponseList = await batchEvents(successRespList);
+      batchedResponseList = batchEvents(successRespList);
     } else {
-      batchedResponseList = await legacyBatchEvents(successRespList);
+      batchedResponseList = legacyBatchEvents(successRespList);
     }
   }
   return [...batchedResponseList, ...errorRespList];
